refactor: generalize template simulation agents
