+++ src/template-validation/simulation-agent-runtime.ts
